Add imagelist.txt files to existing CAP charts
These files help to mirror all required images to a local registry
to reduce downtime during upgrades, or to allow for air-gapped
installation.

The minibroker and nginx-ingress charts are not updated by this commit:

minibroker currently only has a minibroker:latest image in the registry.

nginx-ingress images are in a different organization (sles12 instead of
cap), and also don't use the kube.registry.hostname and kube.organization
settings in values.yaml to specify the local registry.

[CAP-359]